New Q Hotfix for angle and torsion check routine
A new fix was needed for the sanity check routine after reading in angles, torsions and impropers. The previous fix could still lead to memory corruption conditions if the number of changed torsions and the number of broken bonds are not closely matched or if too many broken torsions are missing. This fix should address this issue.