Hotfix for issue 21
This fixes the bug in issue 21 were qcalc would give the wrong energies for q-atom angles.
The cause was not initializing the deg2rad conversion in qatom_startup and a stupid typo in thew energy calc.
The new fix will now properly initialize qatom_startup.
An update for the issue is also presented with the new values.
I also fixed the output format.

Apologies

Paul