Fix for cryptic divide by zero error in qprep5
This commit fixes the divide by zero error in qprep5 if the user tried to create a topology without reading a library that contained information about solvent molecules (e.g. for gas phase simulations).
Thecause was that new versions are prepared to use solvent with more than three atoms and read this information from the library.
If no library was found, this value was set to zero and caused the above error and crash while trying to determine the number of solvent molecules in a PDB file.
This fix adds the check for a loaded solvent, and sets the value for the number of read in solvent molecules to zero if no solvent is loaded.
Sorry for any troubles.
/Paul