Cosmetic fix in qcalc output.
Descriptions for QTorsion calculations were written to stdout instead of
the variable desc. Since writing the string to desc crashes qcalc5 I'm
assuming that this was actually done on purpose as a 'fix'.

I fixed the actual issue of the desc character array being only 60 chars long and
the string 61.